# Use cached etcd artifacts for build