I'm frequently running into performance issues with (re-)building the planet package. This change cuts the time almost in half (in good case - with docker cache and no system package updates).

 * Enable optional caching for go build (off by default), can be enabled with `GOCACHE_ENABLED=yes make ...`
 * Use cached artifacts when building etcd targets

## Testing
```
$ make clean production

# This branch
real	3m20,738s
user	1m15,559s
sys	0m7,153s

# subsequent run
real	2m19,836s
user	1m12,291s
sys	0m4,958s

# master (671409bfdc4852ded8bb04f2669d4636a91269a6)
real	4m11,317s
user	1m15,975s
sys	0m4,885s

# subsequent run
real	3m57,880s
user	1m13,773s
sys	0m4,900s
```